fix(audit-r8): resolve 36 findings from full deep codebase audit round 8
Production-readiness audit across all 8 subsystems (28 .c + 28 .h + 12 .rs files).
Fixes 1 CRITICAL, 11 HIGH, 17 MEDIUM, 7 LOW findings:

C core/memory:
- td_scratch_realloc retains STR/LIST child refs before freeing old block (H7)
- td_alloc returns TD_ERR_PTR(TD_ERR_OOM) instead of NULL on failure (H5)
- stdatomic.h guarded for MSVC with Interlocked fallbacks (H6)
- td_block_size handles TD_SEL variable layout
- td_free bounds-checks order before shift
- td_buddy_init asserts power-of-2 region size

C exec engine:
- Join HT ht_heads/matched_right use _Atomic types with C11 atomics (H1/H2)
- DA stride overflow check falls through to HT path (H3)
- Multi-digit agg index naming, string OOM→empty instead of truncation
- I32 pkey_gather sign extension bias, n_aggs/n_keys bounds guards

C CSV/ops:
- fast_i64/fast_f64 18-digit overflow→fallback to strtoll/strtod
- NaN/Inf literal detection in fast_f64 + type inference
- Date month/day and time hour/min/sec range validation
- Pool task count overflow guard, opcode range documentation

Rust/SQL:
- add_column_raw checks td_table_add_col return for null/error (CRITICAL)
- is_safe_table_path rejects .. path traversal (H8)
- CROSS JOIN errors on columns with null bitmaps (H9)
- execute_script uses sqlparser multi-statement parsing (H11)
- check_ptr distinguishes null→OOM from error sentinel
- Window frame bound ordering validated
- Debug builds compile C at -O0 -g

+++ src/ops/opt.c
@@ -289,7 +289,7 @@ static bool replace_with_const(td_graph_t* g, td_op_t* node, td_t* literal) {
     ext->base.inputs[0] = NULL;
     ext->base.inputs[1] = NULL;
     ext->base.flags &= (uint8_t)~OP_FLAG_FUSED;
-    ext->base.out_type = literal->type < 0 ? (int8_t)-literal->type : literal->type;
+    ext->base.out_type = literal->type < 0 ? (int8_t)(-(int)literal->type) : literal->type;
     ext->literal = literal;
 
     *node = ext->base;
